[FEAT] 전체 기능 검토 완료

+++ client/src/Page/Main.js
@@ -3,8 +3,12 @@ import * as M from "../Page/Styled/StyledComponent";
 
 const Main = () => {
     const handleLogout = () => {
+    if (localStorage.getItem("accessToken")) {
         localStorage.clear();
         alert("로그아웃 되었습니다.")
+    } else {
+        alert("로그아웃 상태입니다.")
+    }
     }
 
     return (
